8602: move answer populating to its own class
